Orgnize client components
