+++ client/src/components/users/usersLists/Followers.js
@@ -1,8 +1,8 @@
 import React, { useContext, useEffect, useState } from "react";
-import { followUser, unfollowUser } from "../helpers/userHelpers";
-import UserContext from "./layout/UserContext";
-import LoaderContext from "./context/LoaderContext";
-import UserCard from "./UserCard";
+import { followUser, unfollowUser } from "../../../helpers/userHelpers";
+import UserContext from "../../context/UserContext";
+import LoaderContext from "../../context/LoaderContext";
+import UserCard from "../UserCard";
 
 const Followers = ({ follower, getUserInfo }) => {
   const [isCurrentUserFollow, setIsCurrentUserFollow] = useState(false);
